feat(robotframework): add support for robotframework 4.x
