+++ src/oxygen/__init__.py
@@ -1,10 +1,6 @@
+from .version import VERSION
 from .base_handler import BaseHandler
 from .oxygen import listener, OxygenLibrary
-from .robot_interface import RobotInterface
-from .version import VERSION
 
-__all__ = ['BaseHandler', 'listener', 'OxygenLibrary', 'RobotInterface']
+__all__ = ['BaseHandler', 'listener', 'OxygenLibrary']
 __version__ = VERSION
-
-
-
